refactor: Pass logger explicitly
This prevents HTTP server startup logs from showing up in the test
output, since we're able to pass a silent logger.
