+++ test/routes/404.test.ts
@@ -4,6 +4,7 @@ import { MemoryAdapter } from 'fs-adapters'
 import assert from 'node:assert'
 import { HttpStatus } from 'omniwheel'
 import { FastifyInstance } from 'fastify'
+import winston from 'winston'
 
 const route = '/does-not-exist'
 
@@ -13,7 +14,7 @@ describe(`route: ${route}`, function () {
 
   it('returns a 404 response', async function () {
     const cache = new Cache(new MemoryAdapter())
-    fastify = await startServer(cache, {})
+    fastify = await startServer(winston.createLogger({ silent: true }), cache, {})
     const response = await fastify.inject({ path: route })
     assert.strictEqual(response.statusCode, HttpStatus.NOT_FOUND)
     assert.deepStrictEqual(response.json(), {
