+++ test/routes/meta/legend.test.ts
@@ -5,6 +5,7 @@ import assert from 'node:assert'
 import { HttpStatus } from 'omniwheel'
 import { legend } from 'ka-mensa-fetch'
 import { FastifyInstance } from 'fastify'
+import winston from 'winston'
 
 const route = '/meta/legend'
 
@@ -14,7 +15,7 @@ describe(`route: ${route}`, function () {
 
   it('returns expected data', async function () {
     const cache = new Cache(new MemoryAdapter())
-    fastify = await startServer(cache, {})
+    fastify = await startServer(winston.createLogger({ silent: true }), cache, {})
     const response = await fastify.inject({ path: route })
     assert.strictEqual(response.statusCode, HttpStatus.OK)
     assert.deepStrictEqual(response.json(), {
